+++ src/entrypoints/background/index.ts
@@ -4,10 +4,13 @@ import { browser, type Browser } from "wxt/browser";
 // Import Types
 import { type ExtensionMessage } from "@/assets/js/types/";
 
-// Import from Script
-import { define }         from "@/assets/js/define";  // メッセージング経由で define を受け取る場合、構造化複製不可能な型が含まれていると送信エラーが発生する為、事前にインポートする@2025/09/13
-import { logging }        from "@/assets/js/logging";
-import { handleOpenURLs } from "./js/openUrlsHandler";
+// Import Module
+import { define }           from "@/assets/js/define";  // メッセージング経由で define を受け取る場合、構造化複製不可能な型が含まれていると送信エラーが発生する為、事前にインポートする
+import { logging }          from "@/assets/js/logging";
+import { initializeConfig } from "@/assets/js/initializeConfig";
+import { handleOpenURLs }   from "./js/openUrlsHandler";
+import { countManager }     from "./js/CountManager";
+import { badgeController }  from "./js/BadgeController";
 
 export default defineBackground({
 	// Set manifest options
@@ -20,6 +23,47 @@ export default defineBackground({
 
 function main() {
 	browser.runtime.onMessage.addListener(eventOnMessage);
+
+	initializeBadgeCounter();
+}
+
+/**
+ * バッジカウンター機能間の連携設定
+ */
+function initializeBadgeCounter() {
+	console.log("Initialize, Badge Counter setting.");
+
+	// カウント数の変更を監視し、バッジのテキストを更新
+	countManager.subscribe((newCount) => {
+		badgeController.updateText(String(newCount));
+	});
+
+	// ストレージの変更を監視し、バッジの色を更新
+	browser.storage.onChanged.addListener(handleStorageOnChanged);
+}
+
+/**
+ * `browser.storage.onChanged` イベントのリスナーとして、ストレージの変更を処理。
+ * 特に、`local` ストレージエリアの `config` オブジェクトの変更を監視し、
+ * バッジの色を更新するために `badgeController.updateColor` を呼び出す。
+ * @param   {object} changes  - 変更されたストレージアイテムのオブジェクト。`key` はストレージアイテムの名前、値は `StorageChange` オブジェクト
+ * @param   {string} areaName - 変更が発生したストレージエリアの名前（例: "local", "sync"）
+ * @returns {void}
+ */
+async function handleStorageOnChanged(changes: { [key: string]: Browser.storage.StorageChange }, areaName: string): Promise<void> {
+	try {
+		if (areaName === "local" && changes.config) {
+			// initializeConfig を経由して最新の設定を取得。`initializeConfig` を用いて値の検証と移行処理を行う
+			// save: false を渡し、onChanged リスナー内で再度保存処理が走らないようにする
+			const { config: updatedConfig } = await initializeConfig(changes.config.newValue, false);
+
+			if (updatedConfig && updatedConfig.Badge) {
+				badgeController.updateColor(updatedConfig.Badge);
+			}
+		}
+	} catch (error) {
+		console.error("Error in browser.storage.onChanged listener:", error);
+	}
 }
 
 /**